#12 상세화면 보강 및 웹드라마 재생 화면 구축

+++ app/src/main/java/com/leisurely/people/enjoyd/data/remote/data/response/DramasSlugGetResponse.kt
@@ -18,7 +18,7 @@ data class DramasSlugGetResponse(
     val title: String,
 
     @SerialName("avg_rating")
-    val avgRating: Int,
+    val avgRating: Double,
 
     val summary: String,
